File crud authen check

+++ src/main/java/com/ntv/ntvcons_backend/entities/PostModels/ShowPostModel.java
@@ -1,6 +1,6 @@
 package com.ntv.ntvcons_backend.entities.PostModels;
 
-import com.ntv.ntvcons_backend.Enum.Status;
+import com.ntv.ntvcons_backend.constants.Status;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -15,5 +15,5 @@ public class ShowPostModel {
 
     private Long createdBy, updatedBy;
     private LocalDateTime createdAt, updatedAt;
-    private Double totalPage;
+    private Integer totalPage;
 }
